+++ packages/domain-events/src/Domain/Event/FileRemovedEventPayload.ts
@@ -1,0 +1,8 @@
+import { Uuid } from '@standardnotes/common'
+
+export interface FileRemovedEventPayload {
+  userUuid: Uuid
+  fileByteSize: number
+  filePath: string
+  fileName: string
+}