+++ packages/features/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.34.5](https://github.com/standardnotes/snjs/compare/@standardnotes/features@1.34.4...@standardnotes/features@1.34.5) (2022-03-09)
+
+**Note:** Version bump only for package @standardnotes/features
+
+
+
+
+
 ## [1.34.4](https://github.com/standardnotes/snjs/compare/@standardnotes/features@1.34.3...@standardnotes/features@1.34.4) (2022-03-08)
 
 **Note:** Version bump only for package @standardnotes/features