+++ packages/snjs/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.79.0](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.78.0...@standardnotes/snjs@2.79.0) (2022-03-10)
+
+
+### Features
+
+* **filepicker:** multiple file selection ([#644](https://github.com/standardnotes/snjs/issues/644)) ([1bcdaf4](https://github.com/standardnotes/snjs/commit/1bcdaf4d2e05e1280ba8646683be71eebf95ee2d))
+
+
+
+
+
+# [2.78.0](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.77.6...@standardnotes/snjs@2.78.0) (2022-03-09)
+
+
+### Features
+
+* export file operation type ([#643](https://github.com/standardnotes/snjs/issues/643)) ([ff5f136](https://github.com/standardnotes/snjs/commit/ff5f136655a8089a47c7eaa04e1e13e58852c93f))
+
+
+
+
+
+## [2.77.6](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.77.5...@standardnotes/snjs@2.77.6) (2022-03-09)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+
+
+
+
+## [2.77.5](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.77.4...@standardnotes/snjs@2.77.5) (2022-03-09)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+
+
+
+
+## [2.77.4](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.77.3...@standardnotes/snjs@2.77.4) (2022-03-09)
+
+**Note:** Version bump only for package @standardnotes/snjs
+
+
+
+
+
 ## [2.77.3](https://github.com/standardnotes/snjs/compare/@standardnotes/snjs@2.77.2...@standardnotes/snjs@2.77.3) (2022-03-09)